requirements.txt
amqp==5.1.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
black==23.3.0
boto==2.49.0
boto3==1.26.131
botocore==1.29.131
celery==5.2.7
certifi==2023.5.7
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.2
django-cors-headers==4.0.0
django-debug-toolbar==3.2.1
django-extensions==3.2.1
django-redis==5.2.0
django-storages==1.13.2
djangochannelsrestframework==1.1.0
djangorestframework==3.14.0
djangorestframework-csv==2.1.1
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
drf-extra-fields==3.7.0
exceptiongroup==1.1.3
fastapi==0.104.1
filetype==1.2.0
gunicorn==20.1.0
h11==0.14.0
idna==3.4
jmespath==1.0.1
kombu==5.2.4
mangum==0.17.0
msgpack==1.0.5
mypy-extensions==1.0.0
mysqlclient==2.1.1
num2words==0.5.12
packaging==23.1
pathspec==0.11.1
pdfkit==1.0.0
pi==0.1.2
Pillow==9.5.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
pydantic==2.5.1
pydantic_core==2.14.3
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
razorpay==1.3.0
redis==4.5.4
requests==2.30.0
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
sqlparse==0.4.3
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.8.0
unicodecsv==0.14.1
urllib3==1.26.15
uvicorn==0.24.0
vine==5.0.0
wcwidth==0.2.6
pydantic-settings==2.1.0